+++ Sources/WalletConnectRelay/RPC/RPCMethod.swift
@@ -1,0 +1,5 @@
+protocol RPCMethod {
+    associatedtype Parameters
+    var method: String { get }
+    var params: Parameters { get }
+}
