+++ Sources/WalletConnectRelay/RPC/WalletConnectRPCID.swift
@@ -1,0 +1,11 @@
+import Foundation
+import JSONRPC
+
+struct WalletConnectRPCID: IdentifierGenerator {
+
+    func next() -> RPCID {
+        let timestamp = Int64(Date().timeIntervalSince1970 * 1000) * 1000
+        let random = Int64.random(in: 0..<1000)
+        return .right(Int(timestamp + random))
+    }
+}
